Move types into common
