+++ src/tools/asset-generate/common/types.ts
